Moving code from src/ to src/deprecated/
