+++ src/deprecated/README.md
@@ -9,3 +9,8 @@ of STARRY/limbdark trying out:
   elliptic integrals
 - derivatives of standard elliptic integrals
 - estimating light travel time delays
+
+sn.jl:  Computed the s_n functions needed for starry (a different
+  set of Green's functions is used in limbdark).
+sn_jacobian.jl:  Computed the derivatives of s_n with auto-diff.
+area_triangle.jl:  Computes are of triangle using Kahan method.
