+++ src/deprecated/area_triangle.jl
